requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pillow
pytesseract
SpeechRecognition
pydub
aiofiles
